functions to messageUtils
